+++ conformity/opencv/tests.cpp
@@ -47,6 +47,7 @@ std::vector<test> all_tests = merge_tests({
     rgb2yuv420_tests_get,
     rgb_to_yuv_422_tests_get,
     standalone_lucas_kanade_alg_tests_get,
+    flip_tests_get,
     rotate_tests_get,
     transpose_tests_get,
     add_padding_by_copy_tests_get,